Reorganize and add names to stupidmodel, reduze tile_size, and fix a few bugs
