+++ examples/stupid/stupid01-basic.py
@@ -4,23 +4,13 @@ from agents import Agent, Model, run
 
 
 class Bug(Agent):
-    def draw_color(self):
-        gradient = max(0, 255-255*self.grow_size/10)
-        self.color = (255, gradient, gradient)
-
     def setup(self, model):
         self.size = 8
-        self.grow_size = 1
-        self.draw_color()
+        self.color = (255, 0, 0)
         self.align()
         self.update_current_tile()
 
     def step(self, model):
-        # Eat from the current tile
-        t = self.current_tile()
-        self.grow_size += min(model["max_food_eat"], t.info["food"])
-        t.info["food"] = max(0, t.info["food"]-model["max_food_eat"])
-
         # Find all nearby valid tiles
         nearby_tiles = self.nearby_tiles(-4, -4, 4, 4)
         random.shuffle(nearby_tiles)
@@ -36,35 +26,23 @@ class Bug(Agent):
             self.jump_to(new_t.x*model.width/model.x_tiles,
                          new_t.y*model.height/model.y_tiles)
             self.align()
-        self.draw_color()
 
 
 def setup(model):
     model.reset()
+    model["initial_bugs"] = 100
     people = set([Bug() for i in range(math.floor(model["initial_bugs"]))])
     model.add_agents(people)
-    for t in model.tiles:
-        t.info["food"] = 0.0
-        t.color = (0, 0, 0)
 
 
 def step(model):
     for a in model.agents:
         a.step(model)
-    for t in model.tiles:
-        food_prod = random.random() * model["max_food_prod"]
-        t.info["food"] += food_prod
-        c = min(255, math.floor(t.info["food"] * 255))
-        t.color = (c, c, c)
-    model.update_plots()
 
 
-stupid_model = Model("Dum-dum", 100, 100)
+stupid_model = Model("Basic StupidModel (stupid01)",
+                     100, 100, tile_size=5)
 stupid_model.add_button("setup", setup)
 stupid_model.add_button("step", step)
 stupid_model.add_toggle_button("go", step)
-stupid_model.add_slider("initial_bugs", 10, 300, 100)
-stupid_model.add_slider("max_food_eat", 0.1, 1.0, 1.0)
-stupid_model.add_slider("max_food_prod", 0.01, 0.1, 0.01)
-stupid_model.histogram_bins("grow_size", 0, 10, 5, (0, 0, 0))
 run(stupid_model)
