+++ examples/stupid/stupid08-write-data-to-file.py
@@ -20,6 +20,8 @@ class Bug(Agent):
         t = self.current_tile()
         self.grow_size += min(model["max_food_eat"], t.info["food"])
         t.info["food"] = max(0, t.info["food"]-model["max_food_eat"])
+        if self.grow_size > 100:
+            model["stop"] = True
 
         # Find all nearby valid tiles
         nearby_tiles = self.nearby_tiles(-4, -4, 4, 4)
@@ -40,11 +42,11 @@ class Bug(Agent):
 
 
 def setup(model):
+    global f
+    f = open("stupid.data", "w")
     model.reset()
-    model["max_food_eat"] = 1.0
-    model["max_food_prod"] = 0.01
-    model["initial_bugs"] = 100
     people = set([Bug() for i in range(math.floor(model["initial_bugs"]))])
+    model["stop"] = False
     model.add_agents(people)
     for t in model.tiles:
         t.info["food"] = 0.0
@@ -52,17 +54,41 @@ def setup(model):
 
 
 def step(model):
-    for a in model.agents:
-        a.step(model)
-    for t in model.tiles:
-        food_prod = random.random() * model["max_food_prod"]
-        t.info["food"] += food_prod
-        c = min(255, math.floor(t.info["food"]/10 * 255))
-        t.color = (c, c, c)
+    global f
+    if not model["stop"]:
+        bug_min = None
+        bug_mean = 0
+        bug_max = None
+        for a in model.agents:
+            a.step(model)
+            if not bug_min or bug_min > a.grow_size:
+                bug_min = a.grow_size
+            if not bug_max or bug_max < a.grow_size:
+                bug_max = a.grow_size
+            bug_mean += a.grow_size
+        bug_mean /= model["initial_bugs"]
+        f.write(str(bug_min) + " " + str(bug_mean) + " " + str(bug_max) + "\n")
+        f.flush() # Flush is necessary as long as we can't call f.close()
+                  # when the user exits the program
+
+        for t in model.tiles:
+            food_prod = random.random() * model["max_food_prod"]
+            t.info["food"] += food_prod
+            c = min(255, math.floor(t.info["food"] * 255))
+            t.color = (c, c, c)
+        model.update_plots()
 
 
-stupid_model = Model("Dum-dum", 100, 100)
+stupid_model = Model("StupidModel w. file output (stupid08)",
+                     100, 100, tile_size=5)
 stupid_model.add_button("setup", setup)
 stupid_model.add_button("step", step)
 stupid_model.add_toggle_button("go", step)
+stupid_model.add_controller_row()
+stupid_model.add_slider("initial_bugs", 10, 300, 100)
+stupid_model.add_controller_row()
+stupid_model.add_slider("max_food_eat", 0.1, 1.0, 1.0)
+stupid_model.add_controller_row()
+stupid_model.add_slider("max_food_prod", 0.01, 0.1, 0.01)
+stupid_model.histogram_bins("grow_size", 0, 10, 5, (0, 0, 0))
 run(stupid_model)
