+++ examples/stupid/stupid11-ranked-neighbour-tiles.py
@@ -31,12 +31,15 @@ class Bug(Agent):
             return len(t.get_agents()) == 0
         nearby_tiles = list(filter(is_valid_tile, nearby_tiles))
 
-        # If there is a valid tile, pick the "first" one and jump to it
-        if len(nearby_tiles) > 0:
-            # Just use the first one in the list, it is shuffled anyways
-            new_t = nearby_tiles[0]
-            self.jump_to(new_t.x*model.width/model.x_tiles,
-                         new_t.y*model.height/model.y_tiles)
+        # Move to the best tile
+        best_t = None
+        for new_t in nearby_tiles:
+            if (not best_t or (best_t.info["food"] < new_t.info["food"])):
+                if len(new_t.get_agents()) == 0:
+                    best_t = new_t
+        if best_t:
+            self.jump_to((best_t.x)*model.width/model.x_tiles,
+                         (best_t.y)*model.height/model.y_tiles)
             self.align()
         self.draw_color()
 
@@ -59,7 +62,7 @@ def step(model):
         bug_min = None
         bug_mean = 0
         bug_max = None
-        for a in model.agents:
+        for a in model.agents_ordered("grow_size"):
             a.step(model)
             if not bug_min or bug_min > a.grow_size:
                 bug_min = a.grow_size
@@ -68,6 +71,8 @@ def step(model):
             bug_mean += a.grow_size
         bug_mean /= model["initial_bugs"]
         f.write(str(bug_min) + " " + str(bug_mean) + " " + str(bug_max) + "\n")
+        f.flush() # Flush is necessary as long as we can't call f.close()
+                  # when the user exits the program
 
         for t in model.tiles:
             food_prod = random.random() * model["max_food_prod"]
@@ -77,12 +82,16 @@ def step(model):
         model.update_plots()
 
 
-stupid_model = Model("Dum-dum", 100, 100)
+stupid_model = Model("StupidModel w. ranked neighbour cells (stupid11)",
+                     100, 100, tile_size=5)
 stupid_model.add_button("setup", setup)
 stupid_model.add_button("step", step)
 stupid_model.add_toggle_button("go", step)
+stupid_model.add_controller_row()
 stupid_model.add_slider("initial_bugs", 10, 300, 100)
+stupid_model.add_controller_row()
 stupid_model.add_slider("max_food_eat", 0.1, 1.0, 1.0)
+stupid_model.add_controller_row()
 stupid_model.add_slider("max_food_prod", 0.01, 0.1, 0.01)
 stupid_model.histogram_bins("grow_size", 0, 10, 5, (0, 0, 0))
 run(stupid_model)
