+++ examples/stupid/stupid14-normal-distribution.py
@@ -1,5 +1,6 @@
 import random
 import math
+import numpy as np
 from agents import Agent, Model, run
 
 
@@ -10,7 +11,7 @@ class Bug(Agent):
 
     def setup(self, model):
         self.size = 8
-        self.grow_size = 1
+        self.grow_size = max(0, np.random.normal(0.1, 0.03))
         self.survivalprobability = 95
         model["current_bugs"] += 1
         self.draw_color()
@@ -96,6 +97,9 @@ def step(model):
             bug_mean += a.grow_size
         bug_mean /= model["initial_bugs"]
         f.write(str(bug_min) + " " + str(bug_mean) + " " + str(bug_max) + "\n")
+        f.flush() # Flush is necessary as long as we can't call f.close()
+                  # when the user exits the program
+
 
         for t in model.tiles:
             food_prod = random.random() * model["max_food_prod"]
@@ -106,12 +110,16 @@ def step(model):
         model.remove_destroyed_agents()
 
 
-stupid_model = Model("Dum-dum", 100, 100)
+stupid_model = Model("StupidModel w. gauss distribution of sizes (stupid14)",
+                     100, 100, tile_size=5)
 stupid_model.add_button("setup", setup)
 stupid_model.add_button("step", step)
 stupid_model.add_toggle_button("go", step)
+stupid_model.add_controller_row()
 stupid_model.add_slider("initial_bugs", 10, 300, 100)
+stupid_model.add_controller_row()
 stupid_model.add_slider("max_food_eat", 0.1, 1.0, 1.0)
+stupid_model.add_controller_row()
 stupid_model.add_slider("max_food_prod", 0.01, 0.1, 0.01)
 stupid_model.histogram_bins("grow_size", 0, 10, 5, (0, 0, 0))
 stupid_model.graph("current_bugs", (0, 0, 0))
