+++ examples/stupid/stupid15-reading-habitat-data.py
@@ -77,7 +77,14 @@ def setup(model):
     model["current_bugs"] = model["initial_bugs"]
     model["stop"] = False
     model.add_agents(people)
-    for t in model.tiles:
+    cell_data = open("stupid.cell", "r")
+    for line in cell_data:
+        cell = line.split()
+        x = int(cell[0])
+        y = int(cell[1])
+        prod_rate = float(cell[2])
+        t = model.tiles[y*model.x_tiles+x]
+        t.info["prod"] = prod_rate
         t.info["food"] = 0.0
         t.color = (0, 0, 0)
 
@@ -97,22 +104,27 @@ def step(model):
             bug_mean += a.grow_size
         bug_mean /= model["initial_bugs"]
         f.write(str(bug_min) + " " + str(bug_mean) + " " + str(bug_max) + "\n")
+        f.flush() # Flush is necessary as long as we can't call f.close()
+                  # when the user exits the program
 
         for t in model.tiles:
-            food_prod = random.random() * model["max_food_prod"]
-            t.info["food"] += food_prod
+            t.info["food"] += t.info["prod"]
             c = min(255, math.floor(t.info["food"] * 255))
             t.color = (c, c, c)
         model.update_plots()
         model.remove_destroyed_agents()
 
 
-stupid_model = Model("Dum-dum", 100, 100)
+stupid_model = Model("StupidModel reading habitat data from file (stupid15)",
+                     250, 112, tile_size=3)
 stupid_model.add_button("setup", setup)
 stupid_model.add_button("step", step)
 stupid_model.add_toggle_button("go", step)
+stupid_model.add_controller_row()
 stupid_model.add_slider("initial_bugs", 10, 300, 100)
+stupid_model.add_controller_row()
 stupid_model.add_slider("max_food_eat", 0.1, 1.0, 1.0)
+stupid_model.add_controller_row()
 stupid_model.add_slider("max_food_prod", 0.01, 0.1, 0.01)
 stupid_model.histogram_bins("grow_size", 0, 10, 5, (0, 0, 0))
 stupid_model.graph("current_bugs", (0, 0, 0))
